Docs addon refactor
