+++ addons/docs/README.md
@@ -1,0 +1,3 @@
+# Storybook Docs
+
+Living documentation for your components.
