+++ addons/docs/register.js
@@ -1,0 +1,1 @@
+require('./dist/register.js');
