+++ addons/docs/src/shared.ts
@@ -1,0 +1,3 @@
+export const ADDON_ID = 'storybooks/docs';
+export const PANEL_ID = `${ADDON_ID}/panel`;
+export const PARAM_KEY = `docs`;
