+++ addons/docs/src/typings.d.ts
@@ -1,0 +1,5 @@
+// todo the following packages need definition files or a TS migration
+declare module '@storybook/components';
+declare module 'global';
+declare module 'react-inspector';
+declare module 'telejson';
