+++ examples/cra-mdx/.storybook/addons.js
@@ -1,3 +1,3 @@
 import '@storybook/addon-actions/register';
 import '@storybook/addon-links/register';
-import './register';
+import '@storybook/addon-docs/register';
