+++ examples/cra-mdx/.storybook/webpack.config.js
@@ -1,4 +1,4 @@
-const createCompiler = require('./storybook-mdx-compiler-plugin');
+const createCompiler = require('@storybook/addon-docs/mdx-compiler-plugin');
 
 module.exports = async ({ config }) => {
   config.module.rules.push({
