+++ examples/cra-mdx/package.json
@@ -24,19 +24,16 @@
   ],
   "devDependencies": {
     "@babel/core": "^7.4.0",
-    "@babel/generator": "^7.4.0",
-    "@babel/parser": "^7.4.2",
     "@mdx-js/loader": "^0.20.3",
     "@mdx-js/mdx": "^0.20.3",
     "@storybook/addon-actions": "5.1.0-alpha.20",
     "@storybook/addon-links": "5.1.0-alpha.20",
+    "@storybook/addon-docs": "5.1.0-alpha.20",
     "@storybook/addons": "5.1.0-alpha.20",
     "@storybook/api": "5.1.0-alpha.20",
     "@storybook/components": "5.1.0-alpha.20",
     "@storybook/react": "5.1.0-alpha.20",
     "babel-loader": "^8.0.5",
-    "lodash.camelcase": "^4.3.0",
-    "prop-types": "^15.7.2",
-    "global": "^4.3.2"
+    "prop-types": "^15.7.2"
   }
 }
