+++ examples/vue-kitchen-sink/.storybook/addons.js
@@ -5,4 +5,4 @@ import '@storybook/addon-knobs/register';
 import '@storybook/addon-viewport/register';
 import '@storybook/addon-options/register';
 import '@storybook/addon-backgrounds/register';
-import './register';
+import '@storybook/addon-docs/register';
