+++ examples/vue-kitchen-sink/.storybook/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const createCompiler = require('./storybook-mdx-compiler-plugin');
+const createCompiler = require('@storybook/addon-docs/mdx-compiler-plugin');
 
 module.exports = async ({ config }) => {
   config.module.rules.push({
