+++ examples/vue-kitchen-sink/package.json
@@ -20,6 +20,7 @@
     "@storybook/addon-centered": "5.1.0-alpha.20",
     "@storybook/addon-knobs": "5.1.0-alpha.20",
     "@storybook/addon-links": "5.1.0-alpha.20",
+    "@storybook/addon-docs": "5.1.0-alpha.20",
     "@storybook/addon-notes": "5.1.0-alpha.20",
     "@storybook/addon-options": "5.1.0-alpha.20",
     "@storybook/addon-storyshots": "5.1.0-alpha.20",
@@ -37,14 +38,8 @@
     "vue-loader": "^15.7.0",
     "webpack": "^4.28.0",
     "webpack-dev-server": "^3.2.0",
-    "lodash.camelcase": "^4.3.0",
     "prop-types": "^15.7.2",
-    "global": "^4.3.2",
     "@mdx-js/loader": "^0.20.3",
-    "@mdx-js/mdx": "^0.20.3",
-    "@babel/generator": "^7.4.0",
-    "@babel/parser": "^7.4.2",
-    "react": "^16.8.5",
-    "react-dom": "^16.8.5"
+    "@mdx-js/mdx": "^0.20.3"
   }
 }
